+++ framework/meta/src/folder_structure/mod.rs
@@ -1,5 +1,7 @@
+mod pretty_print;
 mod relevant_directory;
 mod version_req;
 
+pub use pretty_print::*;
 pub use relevant_directory::*;
 pub use version_req::*;